moves wallet util logic out of wallet module
creates separate utils module within the wallet and submodules for asset and
transaction utils

moves getAssetStatus, getTransactionStatus, and getTransactionType out of the
wallet and into utils
