+++ ironfish/src/rpc/routes/wallet/getAssets.ts
@@ -3,6 +3,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import * as yup from 'yup'
 import { CurrencyUtils } from '../../../utils'
+import { getAssetStatus } from '../../../wallet'
 import { ApiNamespace } from '../namespaces'
 import { routes } from '../router'
 import { AssertHasRpcContext } from '../rpcContext'
@@ -31,7 +32,7 @@ routes.register<typeof GetAssetsRequestSchema, GetAssetsResponse>(
   `${ApiNamespace.wallet}/getAssets`,
   GetAssetsRequestSchema,
   async (request, node): Promise<void> => {
-    AssertHasRpcContext(request, node, 'wallet', 'assetsVerifier')
+    AssertHasRpcContext(request, node, 'wallet', 'assetsVerifier', 'config')
 
     const account = getAccount(node.wallet, request.data.account)
 
@@ -40,6 +41,8 @@ routes.register<typeof GetAssetsRequestSchema, GetAssetsResponse>(
         break
       }
 
+      const confirmations = request.data.confirmations ?? node.config.get('confirmations')
+
       request.stream({
         id: asset.id.toString('hex'),
         metadata: asset.metadata.toString('hex'),
@@ -47,9 +50,7 @@ routes.register<typeof GetAssetsRequestSchema, GetAssetsResponse>(
         creator: asset.creator.toString('hex'),
         owner: asset.owner.toString('hex'),
         nonce: asset.nonce,
-        status: await node.wallet.getAssetStatus(account, asset, {
-          confirmations: request.data.confirmations,
-        }),
+        status: await getAssetStatus(account, asset, confirmations),
         supply: asset.supply !== null ? CurrencyUtils.encode(asset.supply) : undefined,
         createdTransactionHash: asset.createdTransactionHash.toString('hex'),
         verification: node.assetsVerifier.verify(asset.id),
