+++ ironfish/src/rpc/routes/wallet/mintAsset.ts
@@ -9,6 +9,7 @@ import {
 import * as yup from 'yup'
 import { Assert } from '../../../assert'
 import { CurrencyUtils, YupUtils } from '../../../utils'
+import { getAssetStatus } from '../../../wallet'
 import { MintAssetOptions } from '../../../wallet/interfaces/mintAssetOptions'
 import { RpcValidationError } from '../../adapters'
 import { ApiNamespace } from '../namespaces'
@@ -135,6 +136,8 @@ routes.register<typeof MintAssetRequestSchema, MintAssetResponse>(
     const transactionValue = await account.getTransaction(transaction.hash())
     Assert.isNotUndefined(transactionValue)
 
+    const confirmations = request.data.confirmations ?? context.config.get('confirmations')
+
     request.end({
       asset: {
         id: asset.id.toString('hex'),
@@ -143,9 +146,7 @@ routes.register<typeof MintAssetRequestSchema, MintAssetResponse>(
         nonce: asset.nonce,
         creator: asset.creator.toString('hex'),
         owner: asset.owner.toString('hex'),
-        status: await context.wallet.getAssetStatus(account, asset, {
-          confirmations: request.data.confirmations,
-        }),
+        status: await getAssetStatus(account, asset, confirmations),
         createdTransactionHash: asset.createdTransactionHash.toString('hex'),
         verification: context.assetsVerifier.verify(mint.asset.id()),
       },
