+++ ironfish/src/wallet/index.ts
@@ -10,3 +10,5 @@ export { Base64JsonEncoder } from './exporter/encoders/base64json'
 export { JsonEncoder } from './exporter/encoders/json'
 export * from './validator'
 export * from './walletdb/walletdb'
+export * from './utils/asset'
+export * from './utils/transaction'
