+++ ironfish/src/wallet/wallet.test.ts
@@ -13,7 +13,6 @@ import {
   createNodeTest,
   useAccountFixture,
   useBlockWithTx,
-  useBurnBlockFixture,
   useMinerBlockFixture,
   useMinersTxFixture,
   useMintBlockFixture,
@@ -21,9 +20,9 @@ import {
   useTxFixture,
 } from '../testUtilities'
 import { AsyncUtils, BufferUtils, ORE_TO_IRON } from '../utils'
-import { Account, TransactionStatus, TransactionType } from '../wallet'
+import { Account, getTransactionStatus, TransactionStatus } from '../wallet'
 import { MaxMemoLengthError } from './errors'
-import { AssetStatus, Wallet } from './wallet'
+import { Wallet } from './wallet'
 
 describe('Wallet', () => {
   const nodeTest = createNodeTest()
@@ -1316,183 +1315,6 @@ describe('Wallet', () => {
     })
   })
 
-  describe('getTransactionStatus', () => {
-    it('should show unconfirmed transactions as unconfirmed', async () => {
-      const { node } = nodeTest
-
-      const accountA = await useAccountFixture(node.wallet, 'a')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-
-      await node.wallet.updateHead()
-
-      const transaction = blockA1.minersFee
-
-      const transactionValue = await accountA.getTransaction(transaction.hash())
-      Assert.isNotUndefined(transactionValue)
-      Assert.isNotNull(transactionValue.sequence)
-
-      const transactionStatus = await node.wallet.getTransactionStatus(
-        accountA,
-        transactionValue,
-        {
-          headSequence: transactionValue.sequence - 1,
-        },
-      )
-
-      expect(transactionStatus).toEqual(TransactionStatus.UNCONFIRMED)
-    })
-
-    it('should show confirmed transactions as confirmed', async () => {
-      const { node } = nodeTest
-
-      const accountA = await useAccountFixture(node.wallet, 'a')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-
-      await node.wallet.updateHead()
-
-      const transaction = blockA1.minersFee
-
-      const transactionValue = await accountA.getTransaction(transaction.hash())
-      Assert.isNotUndefined(transactionValue)
-
-      // Get status as if head of wallet were much later
-      const transactionStatus = await node.wallet.getTransactionStatus(
-        accountA,
-        transactionValue,
-        {
-          headSequence: 100000,
-        },
-      )
-
-      expect(transactionStatus).toEqual(TransactionStatus.CONFIRMED)
-    })
-
-    it('should show pending transactions as pending', async () => {
-      const { node } = nodeTest
-
-      const accountA = await useAccountFixture(node.wallet, 'a')
-      const accountB = await useAccountFixture(node.wallet, 'b')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-
-      await node.wallet.updateHead()
-
-      const transaction = await useTxFixture(node.wallet, accountA, accountB)
-
-      const transactionValue = await accountA.getTransaction(transaction.hash())
-      Assert.isNotUndefined(transactionValue)
-
-      const transactionStatus = await node.wallet.getTransactionStatus(
-        accountA,
-        transactionValue,
-      )
-
-      expect(transactionStatus).toEqual(TransactionStatus.PENDING)
-    })
-
-    it('should show expired transactions as expired', async () => {
-      const { node } = nodeTest
-
-      const accountA = await useAccountFixture(node.wallet, 'a')
-      const accountB = await useAccountFixture(node.wallet, 'b')
-
-      const blockA2 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA2)
-
-      await node.wallet.updateHead()
-
-      const transaction = await useTxFixture(
-        node.wallet,
-        accountA,
-        accountB,
-        undefined,
-        undefined,
-        3,
-      )
-
-      const blockA3 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA3)
-
-      await node.wallet.updateHead()
-
-      const transactionValue = await accountA.getTransaction(transaction.hash())
-      Assert.isNotUndefined(transactionValue)
-
-      const transactionStatus = await node.wallet.getTransactionStatus(
-        accountA,
-        transactionValue,
-      )
-
-      expect(transactionStatus).toEqual(TransactionStatus.EXPIRED)
-    })
-
-    it('should show transactions with 0 expiration as pending', async () => {
-      const { node } = nodeTest
-
-      const accountA = await useAccountFixture(node.wallet, 'a')
-      const accountB = await useAccountFixture(node.wallet, 'b')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-
-      await node.wallet.updateHead()
-
-      const transaction = await useTxFixture(
-        node.wallet,
-        accountA,
-        accountB,
-        undefined,
-        undefined,
-        0,
-      )
-
-      const transactionValue = await accountA.getTransaction(transaction.hash())
-      Assert.isNotUndefined(transactionValue)
-
-      // Get status as if head of wallet were much later
-      const transactionStatus = await node.wallet.getTransactionStatus(
-        accountA,
-        transactionValue,
-        {
-          headSequence: 100000,
-        },
-      )
-
-      expect(transactionStatus).toEqual(TransactionStatus.PENDING)
-    })
-
-    it('should show unknown status if account has no head sequence', async () => {
-      const { node } = nodeTest
-
-      const accountA = await useAccountFixture(node.wallet, 'a')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-
-      await node.wallet.updateHead()
-
-      const transaction = blockA1.minersFee
-
-      const transactionValue = await accountA.getTransaction(transaction.hash())
-      Assert.isNotUndefined(transactionValue)
-      Assert.isNotNull(transactionValue.sequence)
-
-      await nodeTest.wallet.walletDb.saveHead(accountA, null)
-
-      const transactionStatus = await node.wallet.getTransactionStatus(
-        accountA,
-        transactionValue,
-      )
-
-      expect(transactionStatus).toEqual(TransactionStatus.UNKNOWN)
-    })
-  })
-
   describe('rebroadcastTransactions', () => {
     it('should not rebroadcast expired transactions', async () => {
       const { node } = nodeTest
@@ -1523,9 +1345,10 @@ describe('Wallet', () => {
       const transactionValue = await accountA.getTransaction(transaction.hash())
       Assert.isNotUndefined(transactionValue)
 
-      const transactionStatus = await node.wallet.getTransactionStatus(
+      const transactionStatus = await getTransactionStatus(
         accountA,
         transactionValue,
+        node.config.get('confirmations'),
       )
       expect(transactionStatus).toEqual(TransactionStatus.EXPIRED)
 
@@ -2061,48 +1884,6 @@ describe('Wallet', () => {
     })
   })
 
-  describe('getAssetStatus', () => {
-    it('should return the correct status for assets', async () => {
-      const { node } = await nodeTest.createSetup()
-      const account = await useAccountFixture(node.wallet)
-
-      const mined = await useMinerBlockFixture(node.chain, 2, account)
-      await expect(node.chain).toAddBlock(mined)
-      await node.wallet.updateHead()
-
-      const asset = new Asset(account.publicAddress, 'asset', 'metadata')
-      const value = BigInt(10)
-      const mintBlock = await useMintBlockFixture({
-        node,
-        account,
-        asset,
-        value,
-      })
-
-      let assetValue = await node.wallet.walletDb.getAsset(account, asset.id())
-      Assert.isNotUndefined(assetValue)
-
-      // Check status before added to a block
-      expect(await node.wallet.getAssetStatus(account, assetValue)).toEqual(AssetStatus.PENDING)
-
-      // Add to a block and check different confirmation ranges
-      await expect(node.chain).toAddBlock(mintBlock)
-      await node.wallet.updateHead()
-      assetValue = await node.wallet.walletDb.getAsset(account, asset.id())
-      Assert.isNotUndefined(assetValue)
-      expect(await node.wallet.getAssetStatus(account, assetValue)).toEqual(
-        AssetStatus.CONFIRMED,
-      )
-      expect(
-        await node.wallet.getAssetStatus(account, assetValue, { confirmations: 2 }),
-      ).toEqual(AssetStatus.UNCONFIRMED)
-
-      // Remove the head and check status
-      jest.spyOn(account, 'getHead').mockResolvedValueOnce(Promise.resolve(null))
-      expect(await node.wallet.getAssetStatus(account, assetValue)).toEqual(AssetStatus.UNKNOWN)
-    })
-  })
-
   describe('disconnectBlock', () => {
     it('should update transactions in the walletDb with blockHash and sequence null', async () => {
       const { node } = await nodeTest.createSetup()
@@ -2620,153 +2401,6 @@ describe('Wallet', () => {
       Assert.isNotNull(newAccount)
 
       expect(newAccount.scanningEnabled).toBe(true)
-    })
-  })
-
-  describe('getTransactionType', () => {
-    it('should return miner type for minersFee transactions', async () => {
-      const { node } = await nodeTest.createSetup()
-
-      const accountA = await useAccountFixture(node.wallet, 'a')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-      await node.wallet.updateHead()
-
-      const transaction = blockA1.transactions[0]
-
-      const transactionValue = await accountA.getTransaction(transaction.hash())
-
-      Assert.isNotUndefined(transactionValue)
-
-      await expect(node.wallet.getTransactionType(accountA, transactionValue)).resolves.toEqual(
-        TransactionType.MINER,
-      )
-    })
-
-    it('should return send type for outgoing transactions', async () => {
-      const { node } = await nodeTest.createSetup()
-
-      const accountA = await useAccountFixture(node.wallet, 'a')
-      const accountB = await useAccountFixture(node.wallet, 'b')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-      await node.wallet.updateHead()
-
-      const { block: blockA2, transaction } = await useBlockWithTx(node, accountA, accountB)
-      await expect(node.chain).toAddBlock(blockA2)
-      await node.wallet.updateHead()
-
-      const transactionValue = await accountA.getTransaction(transaction.hash())
-
-      Assert.isNotUndefined(transactionValue)
-
-      await expect(node.wallet.getTransactionType(accountA, transactionValue)).resolves.toEqual(
-        TransactionType.SEND,
-      )
-    })
-
-    it('should return send type for mint transactions', async () => {
-      const { node } = await nodeTest.createSetup()
-
-      const account = await useAccountFixture(node.wallet, 'a')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, account, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-      await node.wallet.updateHead()
-
-      const asset = new Asset(account.publicAddress, 'fakeasset', 'metadata')
-      const value = BigInt(10)
-      const mintBlock = await useMintBlockFixture({
-        node,
-        account,
-        asset,
-        value,
-        sequence: 3,
-      })
-      await expect(node.chain).toAddBlock(mintBlock)
-      await node.wallet.updateHead()
-
-      const transaction = mintBlock.transactions.find((tx) => !tx.isMinersFee())
-
-      Assert.isNotUndefined(transaction)
-
-      const transactionValue = await account.getTransaction(transaction.hash())
-
-      Assert.isNotUndefined(transactionValue)
-
-      await expect(node.wallet.getTransactionType(account, transactionValue)).resolves.toEqual(
-        TransactionType.SEND,
-      )
-    })
-
-    it('should return send type for burn transactions', async () => {
-      const { node } = await nodeTest.createSetup()
-
-      const account = await useAccountFixture(node.wallet, 'a')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, account, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-      await node.wallet.updateHead()
-
-      const asset = new Asset(account.publicAddress, 'fakeasset', 'metadata')
-      const value = BigInt(10)
-      const mintBlock = await useMintBlockFixture({
-        node,
-        account,
-        asset,
-        value,
-        sequence: 3,
-      })
-      await expect(node.chain).toAddBlock(mintBlock)
-      await node.wallet.updateHead()
-
-      const burnValue = BigInt(2)
-      const burnBlock = await useBurnBlockFixture({
-        node,
-        account,
-        asset,
-        value: burnValue,
-        sequence: 4,
-      })
-      await expect(node.chain).toAddBlock(burnBlock)
-      await node.wallet.updateHead()
-
-      const transaction = burnBlock.transactions.find((tx) => !tx.isMinersFee())
-
-      Assert.isNotUndefined(transaction)
-
-      const transactionValue = await account.getTransaction(transaction.hash())
-
-      Assert.isNotUndefined(transactionValue)
-
-      await expect(node.wallet.getTransactionType(account, transactionValue)).resolves.toEqual(
-        TransactionType.SEND,
-      )
-    })
-
-    it('should return receive type for incoming transactions', async () => {
-      const { node } = await nodeTest.createSetup()
-
-      const accountA = await useAccountFixture(node.wallet, 'a')
-      const accountB = await useAccountFixture(node.wallet, 'b')
-
-      const blockA1 = await useMinerBlockFixture(node.chain, undefined, accountA, node.wallet)
-      await expect(node.chain).toAddBlock(blockA1)
-      await node.wallet.updateHead()
-
-      const { block: blockA2, transaction } = await useBlockWithTx(node, accountA, accountB)
-      await expect(node.chain).toAddBlock(blockA2)
-      await node.wallet.updateHead()
-
-      const transactionValue = await accountB.getTransaction(transaction.hash())
-
-      Assert.isNotUndefined(transactionValue)
-
-      await expect(node.wallet.getTransactionType(accountB, transactionValue)).resolves.toEqual(
-        TransactionType.RECEIVE,
-      )
     })
   })
